Working lighting

+++ JackedEngine/Scene/Actors/RotatingPosLightActor.cpp
@@ -10,5 +10,5 @@ RotatingPosLightActor::RotatingPosLightActor(ActorInitializer initializer) :
 void RotatingPosLightActor::Tick(double deltaTime) {
 	BaseActor::Tick(deltaTime);
 	pointLightComponent.SetPosition({ radius * static_cast<float>(sin(alpha)), 0, radius * static_cast<float>(cos(alpha)) });
-	alpha = alpha + deltaTime * rotationSpeed;
+	alpha = alpha + static_cast<float>(deltaTime) * rotationSpeed;
 }
